refactor: split UserProfile into multiple files
