+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2022-03-18T16:43:30.158Z\n"
-"PO-Revision-Date: 2022-03-18T16:43:30.158Z\n"
+"POT-Creation-Date: 2022-03-22T11:54:39.146Z\n"
+"PO-Revision-Date: 2022-03-22T11:54:39.146Z\n"
 
 msgid "Yes"
 msgstr "Yes"
@@ -926,6 +926,15 @@ msgstr "Account disabled?"
 
 msgid "Disabled"
 msgstr "Disabled"
+
+msgid "Male"
+msgstr "Male"
+
+msgid "Female"
+msgstr "Female"
+
+msgid "Other"
+msgstr "Other"
 
 msgid "Overview"
 msgstr "Overview"
