+++ src/pages/UserProfile/Field.js
@@ -1,0 +1,17 @@
+import PropTypes from 'prop-types'
+import React from 'react'
+import styles from './Field.module.css'
+
+const Field = ({ label, value }) => (
+    <div className={styles.field}>
+        <div className={styles.label}>{label}</div>
+        <div>{value}</div>
+    </div>
+)
+
+Field.propTypes = {
+    label: PropTypes.string.isRequired,
+    value: PropTypes.node,
+}
+
+export default Field
