+++ src/pages/UserProfile/Section.module.css
@@ -1,0 +1,18 @@
+.section:not(:last-child) {
+    margin-bottom: var(--spacers-dp32);
+}
+
+.header {
+    display: grid;
+    grid-template-columns: auto 1fr;
+    align-items: center;
+    gap: var(--spacers-dp12);
+}
+
+.title {
+    font-size: 20px;
+    font-weight: 500;
+    color: var(--colors-grey800);
+    margin: 0;
+    margin-bottom: var(--spacers-dp4);
+}
