+++ src/pages/UserProfile/index.js
@@ -1,0 +1,2 @@
+import UserProfile from './UserProfile.js'
+export default UserProfile
